+++ src/Symfony/Bundle/FrameworkBundle/Client.php
@@ -69,7 +69,7 @@ class Client extends BaseClient
             return false;
         }
 
-        return $this->kernel->getContainer()->get('profiler')->loadProfileFromResponse($this->response);
+        return $this->kernel->getContainer()->get('profiler')->loadProfileFromResponse($this->originResponse);
     }
 
     /**