+++ src/Symfony/Bundle/SecurityBundle/Tests/Functional/SwitchUserTest.php
@@ -25,7 +25,7 @@ class SwitchUserTest extends WebTestCase
 
         $client->request('GET', '/profile?_switch_user='.$targetUser);
 
-        $this->assertEquals($expectedStatus, $client->getResponse()->getStatusCode());
+        $this->assertEquals($expectedStatus, $client->getOriginResponse()->getStatusCode());
         $this->assertEquals($expectedUser, $client->getProfile()->getCollector('security')->getUser());
     }
 
@@ -36,7 +36,7 @@ class SwitchUserTest extends WebTestCase
         $client->request('GET', '/profile?_switch_user=user_cannot_switch_1');
         $client->request('GET', '/profile?_switch_user=user_cannot_switch_2');
 
-        $this->assertEquals(500, $client->getResponse()->getStatusCode());
+        $this->assertEquals(500, $client->getOriginResponse()->getStatusCode());
         $this->assertEquals('user_cannot_switch_1', $client->getProfile()->getCollector('security')->getUser());
     }
 
@@ -47,7 +47,7 @@ class SwitchUserTest extends WebTestCase
         $client->request('GET', '/profile?_switch_user=user_cannot_switch_1');
         $client->request('GET', '/profile?_switch_user=_exit');
 
-        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+        $this->assertEquals(200, $client->getOriginResponse()->getStatusCode());
         $this->assertEquals('user_can_switch', $client->getProfile()->getCollector('security')->getUser());
     }