+++ src/Symfony/Component/BrowserKit/CHANGELOG.md
@@ -1,6 +1,16 @@
 CHANGELOG
 =========
 
+2.3.0
+-----
+
+ * added `Client::getOriginRequest()` and `Client::getOriginResponse()` to
+   have access to the origin request and response objects
+ * [BC BREAK] The `Symfony\Component\HttpKernel\Client::getRequest()` method now
+   returns a `Symfony\Component\BrowserKit\Request` instance
+ * [BC BREAK] The `Symfony\Component\HttpKernel\Client::request()` method now
+   always returns a `Symfony\Component\BrowserKit\Response` instance
+
 2.1.0
 -----